requirements.txt
black==24.10.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
idna==3.10
inquirerpy==0.3.4
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
mypy-extensions==1.0.0
new-flask-app==1.0.4
packaging==24.1
pathspec==0.12.1
pfzy==0.3.4
platformdirs==4.3.6
prompt_toolkit==3.0.48
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
